Get rid of old tests/utils.py which was poorly importable.
