+++ tests/test_code.py
@@ -9,9 +9,9 @@ Unit tests for general checks of code quality.
 """
 
 import unittest
-import utils
 import os
 import re
+import parlai.core.testing_utils as testing_utils
 
 FILENAME_EXTENSIONS = r'.*\.(rst|py|sh|js|css)$'
 WHITELIST_PHRASES = [
@@ -29,7 +29,7 @@ class TestCopyright(unittest.TestCase):
     """Make sure all files have the right copyright."""
 
     def test_copyright(self):
-        for fn in utils.git_ls_files():
+        for fn in testing_utils.git_ls_files():
             # only check source files
             if not re.match(FILENAME_EXTENSIONS, fn):
                 continue
@@ -59,7 +59,7 @@ class TestCopyright(unittest.TestCase):
 class TestInit(unittest.TestCase):
     """Make sure all python packages have init files."""
     def test_init_everywhere(self):
-        for folder in utils.git_ls_dirs('parlai'):
+        for folder in testing_utils.git_ls_dirs('parlai'):
             if 'mturk' in folder:
                 continue
             self.assertIn(
