+++ tests/test_new_tasks.py
@@ -9,9 +9,10 @@ import unittest
 
 import parlai.core.teachers as teach_module
 from parlai.scripts.verify_data import verify, setup_args
-from utils import git_changed_files
+import parlai.core.testing_utils as testing_utils
 
 KEYS = ['missing_text', 'missing_labels', 'empty_label_candidates']
+BASE_TEACHERS = dir(teach_module) + ['PytorchDataTeacher']
 
 
 class TestNewTasks(unittest.TestCase):
@@ -20,7 +21,7 @@ class TestNewTasks(unittest.TestCase):
     def test_verify_data(self):
         parser = setup_args()
         opt = parser.parse_args(print_args=False)
-        changed_files = git_changed_files()
+        changed_files = testing_utils.git_changed_files()
         changed_task_files = []
         for file in changed_files:
             if (
@@ -37,13 +38,16 @@ class TestNewTasks(unittest.TestCase):
             task = file.split('/')[-2]
             module_name = "%s.tasks.%s.agents" % ('parlai', task)
             task_module = importlib.import_module(module_name)
-            base_teachers = dir(teach_module) + ['PytorchDataTeacher']
-            subtasks = [':'.join([task, x]) for x in dir(task_module) if
-                        ('teacher' in x.lower() and x not in base_teachers)]
+            subtasks = [
+                ':'.join([task, x])
+                for x in dir(task_module)
+                if ('teacher' in x.lower() and x not in BASE_TEACHERS)
+            ]
 
             for subt in subtasks:
                 opt['task'] = subt
-                text, log = verify(opt, print_parser=False)
+                with testing_utils.capture_output() as _:
+                    text, log = verify(opt, print_parser=False)
                 for key in KEYS:
                     self.assertEqual(
                         log[key],
